Introduce a common class for handling price formatting
