+++ app/src/main/java/com/hicham/wcstoreapp/models/Product.kt
@@ -7,7 +7,7 @@ data class Product(
     val id: Long,
     val name: String,
     val images: List<String>,
-    val prices: Prices,
+    val prices: ProductPrices,
     val shortDescription: String,
     val description: String
 )
